# swagger/swagger.yaml
basePath: /
definitions:
  agentpool.ChatReceipt:
    properties:
      message_id:
        type: string
      question_id:
        type: string
      status:
        type: string
    type: object
...
  localai.AgentAnswerRequest:
    properties:
      question_id:
        type: string
      selected:
        items:
          type: string
        type: array
      text:
        type: string
    type: object
  localai.AgentChatRequest:
    properties:
      conversation_id:
        type: string
      message:
        type: string
    type: object
  localai.AgentInteractionStatus:
    properties:
      status:
        type: string
    type: object
  localai.AgentPlanDecisionRequest:
    properties:
      approved:
        type: boolean
      feedback:
        type: string
      plan_id:
        type: string
      subtasks:
        items:
          type: string
        type: array
    type: object
...
info:
  contact:
    name: LocalAI
    url: https://localai.io
  description: The LocalAI Rest API.
  license:
    name: MIT
    url: https://raw.githubusercontent.com/mudler/LocalAI/master/LICENSE
  title: LocalAI API
  version: 2.0.0
paths:
...
  /api/agents/{name}/answer:
    post:
      consumes:
      - application/json
      parameters:
      - description: Agent name
        in: path
        name: name
        required: true
        type: string
      - description: Target user ID (admin or agent worker only)
        in: query
        name: user_id
        type: string
      - description: Answer
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/localai.AgentAnswerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/localai.AgentInteractionStatus'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
        "501":
          description: Not Implemented
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Answer an agent question
      tags:
      - agents
  /api/agents/{name}/chat:
    post:
      consumes:
      - application/json
      parameters:
      - description: Agent name
        in: path
        name: name
        required: true
        type: string
      - description: Target user ID (admin or agent worker only)
        in: query
        name: user_id
        type: string
      - description: Message
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/localai.AgentChatRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/agentpool.ChatReceipt'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
        "501":
          description: Not Implemented
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Chat with an agent
      tags:
      - agents
  /api/agents/{name}/pending:
    get:
      parameters:
      - description: Agent name
        in: path
        name: name
        required: true
        type: string
      - description: Target user ID (admin or agent worker only)
        in: query
        name: user_id
        type: string
      - description: Conversation ID
        in: query
        name: conversation_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
        "501":
          description: Not Implemented
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get pending agent interactions
      tags:
      - agents
  /api/agents/{name}/plan:
    post:
      consumes:
      - application/json
      parameters:
      - description: Agent name
        in: path
        name: name
        required: true
        type: string
      - description: Target user ID (admin or agent worker only)
        in: query
        name: user_id
        type: string
      - description: Plan decision
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/localai.AgentPlanDecisionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/localai.AgentInteractionStatus'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
        "501":
          description: Not Implemented
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Decide an agent plan
      tags:
      - agents
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"